# ARROW-9216: [C++] Use BitBlockCounter for plain spaced encoding/decoding