Speedup the typical use case which most data are true values, also add null probability
test case.

Signed-off-by: Frank Du <frank.du@intel.com>