@jianxind 2020-06-24
This PR https://github.com/apache/arrow/pull/7532 add the 0.01% benchmark case, I can trigger a benchmark action if 7532 get merged.

Below is the results for 0.01% on my device.
Before
```
BM_PlainEncodingSpacedBoolean/32768/1          74766 ns        74685 ns         9348 bytes_per_second=418.427M/s null_percent=0.01
BM_PlainEncodingSpacedFloat/32768/1            29197 ns        29165 ns        23713 bytes_per_second=4.18549G/s null_percent=0.01
BM_PlainEncodingSpacedDouble/32768/1           31678 ns        31644 ns        22103 bytes_per_second=7.71528G/s null_percent=0.01

BM_PlainDecodingSpacedBoolean/32768/1          34833 ns        34796 ns        20109 bytes_per_second=898.088M/s null_percent=0.01
BM_PlainDecodingSpacedFloat/32768/1            21615 ns        21591 ns        32197 bytes_per_second=5.65367G/s null_percent=0.01
BM_PlainDecodingSpacedDouble/32768/1           24647 ns        24621 ns        28162 bytes_per_second=9.91599G/s null_percent=0.01
```

After
```
BM_PlainEncodingSpacedBoolean/32768/1          51807 ns        51745 ns        13489 bytes_per_second=603.926M/s null_percent=0.01
BM_PlainEncodingSpacedFloat/32768/1             7811 ns         7803 ns        84179 bytes_per_second=15.6441G/s null_percent=0.01
BM_PlainEncodingSpacedDouble/32768/1           13149 ns        13135 ns        49635 bytes_per_second=18.5867G/s null_percent=0.01

BM_PlainDecodingSpacedBoolean/32768/1          18332 ns        18313 ns        38127 bytes_per_second=1.66644G/s null_percent=0.01
BM_PlainDecodingSpacedFloat/32768/1             6634 ns         6627 ns       102921 bytes_per_second=18.4212G/s null_percent=0.01
BM_PlainDecodingSpacedDouble/32768/1           16183 ns        16165 ns        42532 bytes_per_second=15.1027G/s null_percent=0.01
```